refactor(generator): split mixin tests
